lab5_begin

+++ util.c
@@ -16,13 +16,13 @@ void *checked_malloc(int len)
 }
 
 string String(char *s)
-{string p = checked_malloc((int)strlen(s)+1);
+{string p = checked_malloc(strlen(s)+1);
  strcpy(p,s);
  return p;
 }
 
 U_boolList U_BoolList(bool head, U_boolList tail)
-{ U_boolList list = checked_malloc((int)sizeof(*list));
+{ U_boolList list = checked_malloc(sizeof(*list));
   list->head = head;
   list->tail = tail;
   return list;
